@@ -3,5 +3,5 @@
     Glue code to package django projects as yunohost apps.
 """
 
-__version__ = '0.7.1'
+__version__ = '0.8.0'
 __author__ = 'Jens Diemer <git@jensdiemer.de>'
